% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{as.bitvect}
\alias{as.bitvect}
\title{Number to bit vector}
\usage{
as.bitvect(x, length = 32)
}
\arguments{
\item{x}{Input number.}

\item{length}{Length of output bit vector.}
}
\value{
Bit vector.
}
\description{
Convert number to bit vector.
}
\author{
Marvin N. Wright
}
