% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cor.order}
\alias{cor.order}
\title{Order factor levels with correlation approach}
\usage{
cor.order(y, x)
}
\arguments{
\item{y}{Response factor.}

\item{x}{Covariate factor.}
}
\value{
Ordered factor levels
}
\description{
Order factor levels with correlation approach
}
\author{
Marvin N. Wright
}
