% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestClassification.R
\docType{class}
\name{ForestClassification-class}
\alias{ForestClassification-class}
\alias{ForestClassification}
\title{Classification forest class}
\description{
Subclass for classification forest. 
Contains all fields and methods used special for classification forests.
}

