% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestProbability.R
\docType{class}
\name{ForestProbability-class}
\alias{ForestProbability}
\alias{ForestProbability-class}
\title{Probability forest class}
\description{
Subclass for probability forest. 
Contains all fields and methods used special for probability forests.
}

