% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestSurvival.R
\docType{class}
\name{ForestSurvival-class}
\alias{ForestSurvival-class}
\alias{ForestSurvival}
\title{Survival forest class}
\description{
Subclass for survival forest. 
Contains all fields and methods used special for survival forests.
}

