% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{largest.quantile}
\alias{largest.quantile}
\title{Compute largest available quantile}
\usage{
largest.quantile(formula)
}
\arguments{
\item{formula}{Formula for survival model.}
}
\value{
Ordered factor levels
}
\description{
Compute median survival if available or largest quantile available in all strata if median not available.
}
\author{
Marvin N. Wright
}

