% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pc.order}
\alias{pc.order}
\title{Order factor levels with PCA approach}
\usage{
pc.order(y, x)
}
\arguments{
\item{y}{Response factor.}

\item{x}{Covariate factor.}
}
\value{
Ordered factor levels
}
\description{
Order factor levels with PCA approach
}
\references{
Coppersmith, D., Hong, S.J. & Hosking, J.R. (1999) Partitioning Nominal Attributes in Decision Trees. Data Min Knowl Discov 3:197. \url{https://doi.org/10.1023/A:1009869804967}.
}
\author{
Marvin N. Wright
}
