% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reorder.factor.columns}
\alias{reorder.factor.columns}
\title{Reorder factor columns}
\usage{
\method{reorder}{factor.columns}(data)
}
\arguments{
\item{data}{Data with factor columns.}
}
\value{
Data with reordered factor columns.
}
\description{
Reorder factor columns. Use mean for continuous response, class counts for factors and mean survival for survival response.
}
\author{
Marvin N. Wright
}
