% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeProbability.R
\docType{class}
\name{TreeProbability-class}
\alias{TreeProbability-class}
\alias{TreeProbability}
\title{Probability tree class}
\description{
Subclass for probability tree.
Contains all fields and methods used special for probability trees.
}

