% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRegression.R
\docType{class}
\name{TreeRegression-class}
\alias{TreeRegression-class}
\alias{TreeRegression}
\title{Regression tree class}
\description{
Subclass for regression tree.
Contains all fields and methods used special for regression trees.
}

