% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSurvival.R
\docType{class}
\name{TreeSurvival-class}
\alias{TreeSurvival-class}
\alias{TreeSurvival}
\title{Survival tree class}
\description{
Subclass for survival tree.
Contains all fields and methods used special for survival trees.
}

