% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{which.max.random}
\alias{which.max.random}
\title{Index of maximum with random ties}
\usage{
which.max.random(x)
}
\arguments{
\item{x}{Input vector.}
}
\value{
Index of the maximum value.
}
\description{
Find index of maximum, breaking ties at random and ignoring NAs.
}
\author{
Marvin N. Wright
}
