% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{which.min.random}
\alias{which.min.random}
\title{Index of minimum with random ties}
\usage{
which.min.random(x)
}
\arguments{
\item{x}{Input vector.}
}
\value{
Index of the minimum value.
}
\description{
Find index of minimum, breaking ties at random and ignoring NAs.
}
\author{
Marvin N. Wright
}
