% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestRegression.R
\docType{class}
\name{ForestRegression-class}
\alias{ForestRegression-class}
\alias{ForestRegression}
\title{Regression forest class}
\description{
Subclass for regression forest. 
Contains all fields and methods used special for regression forests.
}

