% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeClassification.R
\docType{class}
\name{TreeClassification-class}
\alias{TreeClassification-class}
\alias{TreeClassification}
\title{Classification tree class}
\description{
Subclass for classification tree.
Contains all fields and methods used special for classification trees.
}

